% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLD.R
\name{CLD.test}
\alias{CLD.test}
\title{CLD (Composite Linkage Desequilibrium) based GGI analysis.}
\usage{
CLD.test(Y, G1, G2, n.perm = 1000)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
G1 and G2 arguments (number of individuals).}

\item{G1}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{G2}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{n.perm}{integer strictly superior to 0. This is how much permutations of
Y are performed to compute the pvalue. By default, this is fixed to 1000.}
}
\value{
If the function succeed, it returns the interaction pvalue between G1
 and G2. If the pvalue cannot be computed, an error message is displayed. An
 other error message is shown if the test statistic can't be computed.
}
\description{
\code{CLD.test} performs a Gene-Gene Interaction (GGI) analysis based on the
difference between the covariance matrices of case and control.
}
\details{
The test statistic is computed using a method based on Nagao normalized
Quadratic Distance (NQD). The pvalue is calculated by bootstrap method.
}
\examples{
## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

dta <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)

G1 <- select.snps(dta$snpX, dta$genes.info, "TXNDC5")$snpX
G2 <- select.snps(dta$snpX, dta$genes.info, "DNAH9")$snpX

resp <- system.file("extdata/response.txt", package="GGItest")
Y  <- read.csv(resp, header=FALSE)

## By default, the number of bootstrap replicates is 500
CLD.test(Y, G1, G2)
}
\references{
\enumerate{\item Indika Rajapakse, Michael D. Perlman, Paul J.
 Martin, John A. Hansen, and Charles Kooperberg. Multivariate detection of
 gene-gene interactions. Genetic Epidemiology, 36(6) :622-630, 2012. \item
 Hisao Nagao. On some test criteria for covariance matrix. Ann. Statist.,
 1(4) :700-709, July 1973.}
}
\seealso{
\code{\link{GGI}}
}

