% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI.R
\name{GGI}
\alias{GGI}
\title{Gene-Gene Interaction Analysis Wrapper}
\usage{
GGI(Y, snpX, genes.length = NULL, genes.info = NULL, method = c("PCA.Std",
  "PCA.GenFreq", "CCA", "KCCA", "CLD", "PLSPM", "GBIGM", "minP", "GATES", "tTS",
  "tProd"), ...)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
\code{G1} and \code{G2} (number of individuals).}

\item{snpX}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix}
object. Must have a number of row equal to Y argument's length.}

\item{genes.length}{numeric vector. It is the length (in columns/SNP) of each
gene. Each gene declared is considered contiguous with the one before and
after it. \code{genes.length} can be named if you want the returned matrix
to have dimensions named after those. If no names are given then generic
names are generated following the pattern Gene.n (n being the gene's index)
.}

\item{genes.info}{\emph{(optional)} a data frame. It must have four columns
named \code{Genenames}, \code{SNPnames}, \code{Position} and
\code{Chromosome}. Each row describes a SNP and missing values are not
allowed.}

\item{method}{a string matching onz of the following: PCA.Std, PCA.GenFreq,
CCA, KCCA, CLD, PLSPM, GBIGM, minP, GATES, tTS or tProd. Only one method can
be parsed.}

\item{...}{Other optional arguments to be passed to the functions associated
with the method chosen. See more in elementary methods help.}
}
\value{
\code{GGI} returns a symmetric matrix of size \eqn{G*G} where \eqn{G}
 is the number of genes studied. The general term of the matrix is the
 p-value of the interaction between the two genes.
}
\description{
\code{GGI} performs a Gene-Gene Interaction analysis. Several
 methods are available among which:\itemize{\item PCA-based analysis\item
 CCA-based analyis\item KCCA-based analysis\item CLD-based analysis\item
 PLSPM-based analysis\item GBIGM \item minP test \item GATES test \item tTS
 test \item tProd test}
}
\details{
This function is a wrapper for all GGI analysis methods and drive the
 overall analysis: splitting the dataset in genes matrices and starting
 elementary analysis for each pair of genes.

 \emph{SNP from the same gene must be contiguous. See
 \code{\link{select.snps}}}

 If \code{genes.lenghts} is provided, the values in it are the number of SNP
 from the last gene. Meaning that if \code{genes.length} is as follow: c(20,
 35, 15), then gene 1 will be considered to be the first 20 columns/SNP of
 \code{snpX}, gene 2 will be considered to be the following 35 columns/SNP,
 and so on...

 If \code{genes.info} is provided instead,  then it is only important that
 SNP of a same gene are contiguous. Corresponding columns are automatically
 detected.

 Two work scales are available of which there are specific methods.

 \subsection{Gene's scale}{ When working at gene's scale the available
 methods are:\itemize{ \item Principal Components Analysis method (PCA) - PCA
 is performed on both genes' matrices and resulting principal components are
 used to fit a logit model and test genes' interaction. \item Canonical
 Correlation Analysis (CCA) - The maximum of canonical correlation between
 the two genes is computed for each group (case and control). The difference
 between the two transformed values (Fisher transformation) is used to test
 genes' interaction. \item Kernel Canonical Correlation Analysis (KCCA) -
 This method is the same as CCA but the canonical correlations are computed
 using Kernel method. \item Complete Linkage Desequilibrium (CLD) - Analysis
 based on the difference of the covariance matrices for case and control,
 between the two genes. A method based on Nagao normalized Quadratic Distance
 is used to compute the test statistic. \item Partial Least Square Path
 Modeling (PLSPM) - A network of statistical relations between latent and
 manifest variables is built. The difference between the path coefficients is
 used to compute the test statistic. \item Gene-Based Information Gain Method
 (GBIGM) - Entropies are used to calculate the test statistic and test genes'
 interaction.} }

 \subsection{SNP's scale}{ When working at SNP's scale, a SSI is first
 performed beween each genes combinations and p-values are then calculated
 for GGI using following methods: \itemize{\item Minimum p-value test (minP)
 - The strongest signal is evaluated using multivariate normal distribution.
 \item Gene Association Test using Extended Simes procedure (GATES) - Signals
 are transformed and the strongest one is evaluated. \item Truncated Tail
 Strength test (tTS) - All signals lesser than a set threshold are
 transformed and evaluated. \item Truncated p-value Product test (tProd) -
 Same than tTS with a different signal transformation.} }

 Missing values are rejected and trying to parse an incomplete
 \code{SnpMatrix} object as an argument will result in an error.
}
\examples{
## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")

dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")
dta <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)
resp <- system.file("extdata/response.txt", package="GGItest")

# If a data frame is provided to GGI or one of the *.test function, only the
# column is checked and used.
Y  <- read.csv(resp, header=FALSE)

## By default the PCA-based method is used.
GGI(Y=Y, snpX=dta$snpX, genes.info=dta$genes.info)
}
\seealso{
\code{\link{PCA.Std}}, \code{\link{PCA.GenFreq}},
 \code{\link{CCA.test}}, \code{\link{KCCA.test}}, \code{\link{CLD.test}},
 \code{\link{PLSPM.test}}, \code{\link{GBIGM.test}}, \code{\link{GGI.plot}},
 \code{\link{minP.test}}, \code{\link{gates.test}}, \code{\link{tTS.test}},
 \code{\link{tProd.test}}
}

