% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFile.R
\name{ImportFile}
\alias{ImportFile}
\title{Imports SNPs information from pedfile, PLINK, VCF (4.0) file, or genotypes
imputed by IMPUTE2.}
\usage{
ImportFile(file, pos, pos.sep = "\\t", ...)
}
\arguments{
\item{file}{String containing the path of the file to import.}

\item{pos}{(optional) Path to a csv file, character vector or numeric vector
containing informations for each SNP about chromosome, gene names, snp names
and position.}

\item{pos.sep}{(optional) String to be passed to \code{read.csv} function as
\code{sep} argument. Default is tab (\\t).}

\item{...}{Additionnal arguments to be passed to the reading file function.}
}
\value{
The output is a list of two objects : \code{snpX}, a
 \href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} and \code{genes.info}
 a data frame with 4 columns, and one row per SNP. The columns are
 Chromosome, Genenames, SNPnames and position.
}
\description{
\code{ImportFile} generates a
\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object on the
basis of diallelic object contained in a file, and creates also a data frame
containing information about position of the SNP on the genome.
}
\details{
This function uses the full path to the file to import and read the file with
\code{\link[snpStats]{read.pedfile}}, \code{\link[snpStats]{read.plink}},
\code{\link[GGtools]{vcf2sm}}, or \code{\link[snpStats]{read.impute}},
depending on the file extension (pedfile, plink, vcf or impute2). For a
pedfile, \code{ImportFile} also reads the ".info" file associated (this must
be in the same directory as the ".ped". Similarly, for a PLINK, there must be
3 files with extensions ".bed" (passed as argument \code{file}), ".bim" and
".fam". A VCF file must be with the associated ".tbi" file.

If the file is a vcf file, you must precise two arguments : \itemize{\item
\code{gr} instance of \code{\link[GenomicRanges]{GRanges}.} \item
\code{nmetacol} numeric defining the number of columns used in each record as
locus-level metadata.}

Pos argument is optional. If it's not given, the data frame with position
information is filled with NAs, except for SNP names which are imported from
the column names of the
\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix}, and eventually
positions and chromosomes if the file imported is a pedfile or a PLINK. Else,
the pos argument can be either the path to a csv file, a character vector with
elements of the form chr:position, or a numeric vector with only the
positions. Additionnaly, SNP names can be precised as names of the vector. If
you choose the csv file path, be sure that the columns are named like
following : Chromosome, Genenames, SNPnames, Position. Other writings are
understood, refer to the function itself to know more.
}
\examples{
## Pedfile from this package.
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")

## Information about position of the snps
posi <- system.file("extdata/example.txt", package="GGItest")

## Importation
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

#########
## VCF file from GGtools package.
vref <- system.file("vcf/CEU.exon.2010_09.genotypes.vcf.gz", package="GGtools")
irange <- IRanges::IRanges(10e6,20e6)
gg = GenomicRanges::GRanges(seqnames="1", ranges=irange)
dta <- ImportFile(file=vref, gr=gg, nmetacol=9L)
}

