% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{imputeSnpMatrix}
\alias{imputeSnpMatrix}
\title{Missing values handling for SnpMatrix object}
\usage{
imputeSnpMatrix(snpX, genes.info, on.rem = c("SNP", "ind", "none"),
  quiet = FALSE)
}
\arguments{
\item{snpX}{\code{snpMatrix} object of which SNP are to be removed}

\item{genes.info}{data frame with four columns named \code{Genenames},
\code{SNPnames}, \code{Position} and \code{Chromosome}. Each row describes a
SNP and missing values are not allowed. If \code{genes.info} is provided an
updated version is returned.}

\item{on.rem}{\emph{(optional)} a character string matching one of the
following items: SNP, ind, none. Describes the action taken in case of
remaining missing values. \emph{See details}}

\item{quiet}{\emph{(optional)} a boolean describing wether or not progress bar
should be displayed.}
}
\value{
a list object with two named elements:\describe{ \item{\code{snpX}}{a
 \code{SnpMatrix} object corresponding to input matrix with imputed values.}
 \item{\code{genes.info}}{a data frame object corresponding to an updated
 version of input \code{genes.info} in case SNP had to be removed.}}

 A warning is issued when SNP or individuals had to be removed.
}
\description{
\code{imputeSnpMatrix} is a generic wrapper of
 \code{snp.imputation} and \code{impute.snps} functions from \code{snpStats}
 package. This function mimics a Leave-One-Out process where missing SNP are
 imputed for an individual based on a model trained on all other individuals.
 It is always preferable to use \code{snp.imputation} and \code{impute.snps}
 with settings specific to imputed dataset.
}
\details{
For each individual in the dataset, following process is performed:
 \itemize{\item missing SNP are detected for individual i \item rules are
 imputed for every missing SNP using the whole dataset where individual i is
 removed \item SNP are imputed for individual i} This allows to impute a
 large number of missing SNP but some missing values may remains.

 In that case the choice is left to the user about what action to take: leave
 the dataset as it is, remove all SNP or individuals with missing values.
 Removing all SNP is often more parsimonious than removing individuals and
 allows to get a dataset without any missing values with minimum
 information-loss.
}
\examples{
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

## In this example, genes are loosely scoured but default are much harsher
## conditions
imputed.snps <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)
}
\seealso{
\code{\link{GGI}} \code{\link{snpMatrixScour}}
}

