% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA.Std}
\alias{PCA.GenFreq}
\alias{PCA.Std}
\title{PCA-based GGI analysis.}
\usage{
PCA.Std(Y, G1, G2, threshold = 0.8)

PCA.GenFreq(Y, G1, G2, threshold = 0.8)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
\code{G1} and \code{G2} (number of individuals).}

\item{G1}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{G2}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{threshold}{(optional) numeric comprised in \eqn{[0, 1]} or
\eqn{[0, 100]} intervals. This is how much inertia should be kept for each
PCA and has a direct impact on how many PCA dimensions are kept.}
}
\value{
If the function succeed, the interaction p-value is returned.

 A warning can be issued if the inertia threshold is too high and principal
 components had to be removed.

 An error is returned if too much PCA dimensions had to be removed.
}
\description{
\code{PCA.Std} performs a Gene-Gene Interaction (GGI) analysis
 using a PCA in which dataset is standardized using variables' standard
 deviation. With \code{PCA.GenFreq}, dataset is standardized using the
 standard deviation under Hardy-Weinberg equilibrium.
}
\details{
A Principal Components Analysis is performed on both \code{G1} and
 \code{G2} matrices. With \code{PCA.Std} dataset is standardized using
 variables' standard deviation. Whereas with \code{PCA.GenFreq}, dataset is
 standardized using standard deviation under Hardy-Weinberg equilibrium. Each
 of these two matrices represents a gene to test and is built as follow:
 \itemize{ \item A column is a SNP of said gene. \item A row is an individual
 for which the haplotype and the response variable are known.}

 Principal components are retrieved to describe each dataset with user-set
 inertia percentage.

 Principal components are then used in a logit model fitting process in which
 interaction is tested. If model fitting fails then a PC is removed from the
 pool and a new attempt to fitting is made. The process loops until either
 the model is fitted or there are only two PC for each genes remaining.
 Principal components are removed according to lowest inertia percentage.
}
\section{Functions}{
\itemize{
\item \code{PCA.GenFreq}: Standardization based on Hardy-Weinberg equilirum
}}
\examples{
## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

dta <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)

G1 <- select.snps(dta$snpX, dta$genes.info, "TXNDC5")$snpX
G2 <- select.snps(dta$snpX, dta$genes.info, "DNAH9")$snpX

resp <- system.file("extdata/response.txt", package="GGItest")
Y  <- read.csv(resp, header=FALSE)

## By default, inertia threshold is 80\%
PCA.Std(Y, G1, G2)

## Setting a higher inertia threshold often ends up with removed PCA
## dimensions
PCA.GenFreq(Y, G1, G2, threshold = 1)
}
\references{
Jia Li, Rui Tang, Joanna Biernacka and Maricka de Andrade.
 (2009) Identification of gene-gene interaction using principal components. BMC
 Proceedings, 3 (Suppl. 7): S78
}
\seealso{
\code{\link{GGI}}
}

