% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSPM.R
\name{PLSPM.test}
\alias{PLSPM.test}
\title{PLSPM (Partial Least Squares Path Modelling Analysis) based GGI analysis.}
\usage{
PLSPM.test(Y, G1, G2)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
G1 and G2 arguments (number of individuals).}

\item{G1}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{G2}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}
}
\value{
If the function succeed, it returns the interaction pvalue between G1
 and G2. If the pvalue cannot be computed, an error message is displayed.
}
\description{
\code{PLSPM.test} performs a Gene-Gene Interaction (GGI) analysis based on the
modelisation of a statistical relations network. The aim is to quantify the
connections between the latent and the manifest variables.
}
\details{
To calculate the test statistic for the interaction pvalue, \code{PLSPM.test}
uses \code{\link[plspm]{plspm}} for path coefficients and variances
estimation. The pvalue is obtained by using a permutation method on the
difference between the path coefficients.
}
\examples{
## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

dta <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)

G1 <- select.snps(dta$snpX, dta$genes.info, "bub3")$snpX
G2 <- select.snps(dta$snpX, dta$genes.info, "CA1")$snpX

resp <- system.file("extdata/response.txt", package="GGItest")
Y  <- read.csv(resp, header=FALSE)

## By default, the number of bootstrap replicates is 500
PLSPM.test(Y, G1, G2)
}
\references{
\enumerate{\item Xiaoshuai Zhang, Xiaowei Yang, Zhongshang Yuan,
 Yanxun Liu, Fangyu Li, Bin Peng, Dianwen Zhu, Jinghua Zhao, and Fuzhong Xue.
 A plspm-based test statistic for detecting gene-gene co-association in
 genome-wide association study with case-control design. PLoS ONE, 8(4)
 :e62129, April 2013. \item H. Wold. Estimation of principal components and
 related models by iterative least squares. Academic Press, pages 391-420,
 1966. \item Doerge RW and Churchill GA. Permutation tests for multiple loci
 affecting a quantitative character. Genetics, (142) :285, 1996. \item M.
 Hallin and C. Ley. Permutation tests. Wiley StatsRef : Statistics Reference
 Online, 2014.}
}
\seealso{
\code{\link{GGI}}
}

