% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.snps.R
\name{select.snps}
\alias{select.snps}
\title{Selects SNPs in snpXatrix and associated info data.frame, outputs of
ImportFile.}
\usage{
select.snps(snpX, genes.info, select)
}
\arguments{
\item{snpX}{\href{http://bioconductor.org/packages/snpStats/}{snpXatrix}
object. Given as output of \code{\link{ImportFile}}.}

\item{genes.info}{Data.frame containing informations about snps. For more
details, refer to \code{\link{ImportFile}} help file.}

\item{select}{Numeric or character vector for selecting snps in \code{snpX}
and \code{genes.info}. See details for more information.}
}
\value{
The output is a list of two objects : \code{snpX}, a
 \href{http://bioconductor.org/packages/snpStats/}{snpXatrix} and \code{genes.info}
 a data frame with 4 columns, and one row per SNP selected with
 \code{select}. The columns are Chromosome, Genenames, SNPnames and Position.
 An error message is displayed if the genes of snps selected are not found in
 the either \code{snpX} or \code{genes.info}.
}
\description{
\code{select.snps} allows the user to select snps from an object output of
\code{\link{ImportFile}}. It generates the same object, with the columns of
the \href{http://bioconductor.org/packages/snpStats/}{snpXatrix} and the rows
of the data.frame corresponding to the selected snps.
}
\details{
For a good working of the function, let the column names of the
\code{genes.info} data.frame as they were when outputted from
\code{\link{ImportFile}}.

\code{select} argument can be either : \itemize{\item a numeric vector with
only the column number in the
\href{http://bioconductor.org/packages/snpStats/}{snpXatrix} (or row number
for \code{genes.info}) of each snp selected.\item a character vector with the
names of each snp selected or each gene selected.\item a character vector
which elements are position bounds of genes. Each element of the vector is
either of the form "begin:end", or "chr:begin:end" if you have to precise the
chromosome of the gene.}
}
\examples{

## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

## Selection of the genes DNAH9 and TXNDC5
selec <- select.snps(dta$snpX, dta$genes.info, c("DNAH9","TXNDC5"))

#' ## Selection of the snps from position 101342000 to 101490000 on chromosome 15
selec <- select.snps(dta$snpX, dta$genes.info, c("15:101342000:101490000"))
}

