% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIGM.R
\name{GBIGM.test}
\alias{GBIGM.test}
\title{GBIGM (Gene-based information gain method) for GGI analysis.}
\usage{
GBIGM.test(Y, G1, G2, n.perm = 1000)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
G1 and G2 arguments (number of individuals).}

\item{G1}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{G2}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{n.perm}{integer strictly superior to 0. This is how much permutations
of Y are performed to compute the pvalue. By default, this is fixed to 1000.}
}
\value{
The function returns the interaction pvalue between G1 and G2.
}
\description{
\code{GBIGM.test} performs a Gene-Gene Interaction (GGI) analysis based on the
entropies calculus. It allows detecting non linear interaction between 2
genes.
}
\details{
First, the entropy for each gene \eqn{H(X1)} and \eqn{H(X2)} are calculated.
They are used to compute the conditionnal entropies for each gene
\eqn{H(Y|X1)} and \eqn{H(Y|X2)} and for both \eqn{H(Y|X1;X2)}. Then, the test
statistic can be computed with these datas. Pvalue is calculated using
bootstrap method.
}
\examples{
## Dataset is included in the package
ped <- system.file("extdata/example.ped", package="GGItest")
info <- system.file("extdata/example.info", package="GGItest")
posi <- system.file("extdata/example.txt", package="GGItest")
dta <- ImportFile(file=ped, snps=info, pos=posi, pos.sep="\\t")

dta <- imputeSnpMatrix(dta$snpX, genes.info = dta$genes.info)

G1 <- select.snps(dta$snpX, dta$genes.info, "TXNDC5")$snpX
G2 <- select.snps(dta$snpX, dta$genes.info, "DNAH9")$snpX

resp <- system.file("extdata/response.txt", package="GGItest")
Y  <- read.csv(resp, header=FALSE)

## By default, the number of bootstrap replicates is 500
GBIGM.test(Y, G1, G2)
}
\references{
J. Li, D. Huang, M. Guo, X. Liu, C. Wang, Z. Teng, R. Zhang, Y.
 Jiang, H. Lv, and L. Wang. A gene-based information gain method for
 detecting gene-gene interactions in case-control studies. European Journal
 of Human Genetics, Online : Online, March 2015
}
\seealso{
\code{\link{GGI}}
}

