% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCCA.R
\name{KCCA.test}
\alias{KCCA.test}
\title{KCCA (Kernel Canonical-Correlation Analysis) based GGI analysis.}
\usage{
KCCA.test(Y, G1, G2, n.boot = 500)
}
\arguments{
\item{Y}{numeric or factor vector with two values (most often 0, 1). This is
the response variable and should be of length equal to the number of row of
G1 and G2 arguments (number of individuals).}

\item{G1}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{G2}{\href{http://bioconductor.org/packages/snpStats/}{SnpMatrix} object.
Must have a number of row equal to Y argument's length.}

\item{n.boot}{integer strictly superior to 0. This is the number of bootstrap
replicates for the variances estimation. By default, this is fixed to 500.}
}
\value{
If the function succeed, it returns the interaction pvalue between G1
 and G2. An error message is displayed if the variances used for the test
 statistic calculus could not be estimated. An error message is shown if the
 test statistic could not be computed.
}
\description{
\code{KCCA.test} performs a Gene-Gene Interaction (GGI) analysis based on the
difference of canonical correlations between cases and controls. The "kernel trick" is applied to the canonical correlation to allow the detection of non-linear co-association.
}
\details{
	The test statistic is based on the difference between a Fisher's' transformation of the maximum of the kernelized canonical correlations in cases and controls. To calculate the test statistic for the interaction pvalue, \code{CCA.test} estimates the variance of the Fisher's' transformation of the maximum of the kernelized canonical correlations in cases and controls using a bootstrap method.
	The computation of \code{\link[kernlab]{kcca}.} can be very long.
}
\examples{
	data(gene.pair)
	KCCA.test(Y=gene.pair$Y, G1=gene.pair$G1,G2=gene.pair$G2)
}
\references{
\enumerate{\item Yuan, Z. et al. (2012): Detection for gene-gene co-association via kernel canonical correlation analysis, BMC Genetics, 13, 83.
	\item Larson, N. B. et al. (2013): A kernel regression approach to gene-gene interaction detection for case-control studies, Genetic Epidemiology, 37, 695–703.
}
\seealso{
\code{\link{GGI}}, \code{\link{CCA.test}}
}

