% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.snps.R
\name{print.GGItest}
\alias{print.GGItest}
\title{Printing objects of class \code{"GGItest"}}
\usage{
print.GGItest(x)
}
\arguments{
\item{x}{\\code{GGItest}. Given as output of \code{\link{CCA.test}}.}

}
\value{
the argument x, invisibly, as for all \code{\link{print}} methods.}
\description{
	Printing objects of class \code{"GGItest"}
}
\details{
\code{print.GGItest} provides a customized output of a \code{GGItest} object.
}
\examples{

## Dataset example is included in the package
PCA.Std(Y=example$Y, G1=example$G1,G2=example$G2,threshold=0.7)
print(PCA.Std(Y=example$Y, G1=example$G1,G2=example$G2,threshold=0.7))

}

